grievance-mvp/apps/smtp_graph_bridge/requirements.txt
aiosmtpd==1.4.6
msal==1.34.0
PyYAML==6.0.2
requests==2.32.3
